CIV-10428 Moved files into model folder (Corrected Misplacement)
