+++ src/main/java/uk/gov/hmcts/reform/civil/model/genapplication/finalorder/AppealTypeChoices.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.enums.dq;
+package uk.gov.hmcts.reform.civil.model.genapplication.finalorder;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
