+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/user/JudicialFinalDecisionHandlerTest.java
@@ -11,8 +11,8 @@ import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse
 import uk.gov.hmcts.reform.ccd.client.model.SubmittedCallbackResponse;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.enums.YesOrNo;
-import uk.gov.hmcts.reform.civil.enums.dq.AppealTypeChoiceList;
-import uk.gov.hmcts.reform.civil.enums.dq.AppealTypeChoices;
+import uk.gov.hmcts.reform.civil.model.genapplication.finalorder.AppealTypeChoiceList;
+import uk.gov.hmcts.reform.civil.model.genapplication.finalorder.AppealTypeChoices;
 import uk.gov.hmcts.reform.civil.enums.dq.FinalOrderSelection;
 import uk.gov.hmcts.reform.civil.handler.callback.BaseCallbackHandlerTest;
 import uk.gov.hmcts.reform.civil.model.CaseData;
